Remove unused SetProperty command
